requirements.txt
langchain>=1.0.0
langgraph>=1.0.0
langchain-deepseek>=0.1.0
langchain-huggingface>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
arxiv>=2.1.0
beautifulsoup4>=4.12.0
requests>=2.32.0
